% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2odb.R
\name{list2odb}
\alias{list2odb}
\title{Write ODB file from formatted list}
\usage{
list2odb(list, file, ...)
}
\arguments{
\item{list}{A list imported by \code{\link[=xlsx2list]{xlsx2list()}}.}

\item{file}{Character value indicating the path and/or file name of resulting
ODB file.}

\item{...}{Further arguments (not yet in use).}
}
\value{
An ODB file.
}
\description{
The list produced by \code{\link[=xlsx2list]{xlsx2list()}} will be written in a relational database
as ODB file (for OpenOffice or LibreOffice).
}
