% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odb2df.R
\name{odb2df}
\alias{odb2df}
\title{Import ODB database into a single data frame}
\usage{
odb2df(file, ...)
}
\arguments{
\item{file}{Character value indicating path and/or name of ODB file to be
imported.}

\item{...}{Further arguments (not yet in use).}
}
\value{
A data frame.
}
\description{
ODB files written by \code{\link[=list2odb]{list2odb()}} will be queried into a single data frame.
}
