% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_calc.R
\name{rate_calc}
\alias{rate_calc}
\title{Calculation of growth rates}
\usage{
rate_calc(
  vars = c("height_cm", "bd_mm", "dbh_mm"),
  by = "individual_no",
  time = "record_date",
  data,
  prefix = "r_",
  suffix = "",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{vars}{Character vector indicating the variables in \code{data} that will be
assessed.}

\item{by}{Character value indicating the variable in \code{data} containing the
identity of observed objects.}

\item{time}{Character value indicating the variable in \code{data} containing the
time progress.}

\item{data}{A data frame containing the mentioned variables.}

\item{prefix}{A character value indicating a prefix added to the assessed
variable in the output.}

\item{suffix}{A character value used as suffix.}

\item{na.rm}{A logical value indicating whether missing values in the
assessed variable should be removed before doing the calculation or not.}

\item{...}{Further variables (not yet in use).}
}
\value{
Data frame \code{data} with added variables including calculated rates.
}
\description{
Rates of increase of variables specified in argument \code{vars} will be
calculated in function of the variable \code{time} and the identity of observed
objects specified in argument \code{by}.
}
