% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_growth.R
\name{read_growth}
\alias{read_growth}
\title{Read growth data and produce an ODB file}
\usage{
read_growth(xlsx, odb = ".temp.odb", ...)
}
\arguments{
\item{xlsx}{Character value with the path and/or name of the XLSX input file.}

\item{odb}{Character value indicating the path and/or name of output ODB
file.}

\item{...}{Further arguments (not yet in use).}
}
\value{
A data frame as in \code{\link[=odb2df]{odb2df()}}.
}
\description{
Read row data and export to a single data frame using an ODB file as
intermediating file.

This function applies the sequence \code{\link[=xlsx2list]{xlsx2list()}} -> \code{\link[=list2odb]{list2odb()}} ->
\code{\link[=odb2df]{odb2df()}}.
}
