% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx2list.R
\name{xlsx2list}
\alias{xlsx2list}
\title{Read xlsx book into a list}
\usage{
xlsx2list(file, ...)
}
\arguments{
\item{file}{Character value indicating the path and/or the name of the file
containing the data to be imported.}

\item{...}{Further arguments (not yet used).}
}
\value{
A list with four data frames, namely \code{plots}, \code{species}, \code{individuals}, and
\code{records}.
}
\description{
This function reads and format the data into a list, preparing it for export
into a relational database.
Data frames will be imported from an MS-Excel book by the function
\code{\link[=read.xlsx]{read.xlsx()}}.
}
